requirements.txt
torch
torch-geometric
optuna
wandb
numpy
pandas
scikit-learn
matplotlib
networkx
seaborn
tqdm
jupyter
scipy
PyYAML
python-louvain